% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CONCEPT}
\alias{CONCEPT}
\title{CONCEPT}
\format{
A data frame with 931 rows and 10 variables:
\describe{
\item{\code{concept_id}}{double COLUMN_DESCRIPTION}
\item{\code{concept_name}}{character COLUMN_DESCRIPTION}
\item{\code{domain_id}}{character COLUMN_DESCRIPTION}
\item{\code{vocabulary_id}}{character COLUMN_DESCRIPTION}
\item{\code{concept_class_id}}{character COLUMN_DESCRIPTION}
\item{\code{standard_concept}}{character COLUMN_DESCRIPTION}
\item{\code{concept_code}}{character COLUMN_DESCRIPTION}
\item{\code{valid_start_date}}{double COLUMN_DESCRIPTION}
\item{\code{valid_end_date}}{double COLUMN_DESCRIPTION}
\item{\code{invalid_reason}}{character COLUMN_DESCRIPTION}
}
}
\usage{
CONCEPT
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
