% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CONCEPT_DEPRECATION}
\alias{CONCEPT_DEPRECATION}
\title{CONCEPT_DEPRECATION}
\format{
A data frame with 27026 rows and 4 variables:
\describe{
\item{\code{cd_datetime}}{double COLUMN_DESCRIPTION}
\item{\code{principle_concept_id}}{double COLUMN_DESCRIPTION}
\item{\code{concept_name}}{character COLUMN_DESCRIPTION}
\item{\code{deprecated_concept_id}}{double COLUMN_DESCRIPTION}
}
}
\usage{
CONCEPT_DEPRECATION
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
