% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CONCEPT_SYNONYM}
\alias{CONCEPT_SYNONYM}
\title{CONCEPT_SYNONYM}
\format{
A data frame with 7828 rows and 3 variables:
\describe{
\item{\code{concept_id}}{double COLUMN_DESCRIPTION}
\item{\code{concept_synonym_name}}{character COLUMN_DESCRIPTION}
\item{\code{language_concept_id}}{double COLUMN_DESCRIPTION}
}
}
\usage{
CONCEPT_SYNONYM
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
