% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LINKS_TO_RESOURCES}
\alias{LINKS_TO_RESOURCES}
\title{LINKS_TO_RESOURCES}
\format{
A data frame with 11769 rows and 4 variables:
\describe{
\item{\code{ltr_datetime}}{double COLUMN_DESCRIPTION}
\item{\code{resource_agency}}{character COLUMN_DESCRIPTION}
\item{\code{resource_link}}{character COLUMN_DESCRIPTION}
\item{\code{rn_url}}{character COLUMN_DESCRIPTION}
}
}
\usage{
LINKS_TO_RESOURCES
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
