% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REGISTRY_NUMBERS}
\alias{REGISTRY_NUMBERS}
\title{REGISTRY_NUMBERS}
\format{
A data frame with 2968 rows and 4 variables:
\describe{
\item{\code{rn_datetime}}{double COLUMN_DESCRIPTION}
\item{\code{rn_url}}{character COLUMN_DESCRIPTION}
\item{\code{concept_registry_number_type}}{character COLUMN_DESCRIPTION}
\item{\code{concept_registry_number}}{character COLUMN_DESCRIPTION}
}
}
\usage{
REGISTRY_NUMBERS
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
