% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RN_URL_VALIDITY}
\alias{RN_URL_VALIDITY}
\title{RN_URL_VALIDITY}
\format{
A data frame with 11405 rows and 5 variables:
\describe{
\item{\code{c_datetime}}{double COLUMN_DESCRIPTION}
\item{\code{concept_classification}}{character COLUMN_DESCRIPTION}
\item{\code{rn_url}}{character COLUMN_DESCRIPTION}
\item{\code{rnuv_datetime}}{double COLUMN_DESCRIPTION}
\item{\code{is_404}}{character COLUMN_DESCRIPTION}
}
}
\usage{
RN_URL_VALIDITY
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
