% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NAMES_AND_SYNONYMS}
\alias{NAMES_AND_SYNONYMS}
\title{NAMES_AND_SYNONYMS}
\format{
A data frame with 9245 rows and 4 variables:
\describe{
\item{\code{nas_datetime}}{double COLUMN_DESCRIPTION}
\item{\code{rn_url}}{character COLUMN_DESCRIPTION}
\item{\code{concept_synonym_type}}{character COLUMN_DESCRIPTION}
\item{\code{concept_synonym_name}}{character COLUMN_DESCRIPTION}
}
}
\usage{
NAMES_AND_SYNONYMS
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
